pyproject.toml
[tool.poetry]
name = "pythonproject4"
version = "0.1.0"
description = ""
authors = ["Grigoriy1209 <grigoriyv1209@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
